+++ ft_printf/utils/conversion.c
@@ -6,7 +6,7 @@
 /*   By: jkauker <jkauker@student.42heilbrnn.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/08 21:26:52 by jonask            #+#    #+#             */
-/*   Updated: 2023/10/23 15:12:15 by jkauker          ###   ########.fr       */
+/*   Updated: 2023/10/23 16:14:52 by jkauker          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -76,7 +76,7 @@ void	put_pointer(void *ptr, int *writecount)
 	if (ptr == 0)
 		put_string("0x0", writecount);
 	else
-		put_base((uintptr_t)ptr, 16, 0, writecount);
+		put_base((unsigned long) ptr, 16, 0, writecount);
 }
 
 // void	put_double(double num)
